Refactored organizer autocomplete
